Improved enemy AI
adds exploding AI, also changes how damage is dealt to player (Before there'd be a timer on player, now its on Enemy)
