--- Survival-Game/Assets/Scripts/Enemy/EnemyAI.cs
@@ -1,26 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.AI;
-
-public class EnemyAI : MonoBehaviour
-{
-
-    [SerializeField] private GameObject player;
-    [SerializeField] private PlayerHealth playerHealth;
-    private NavMeshAgent meshAgent;
-
-    void Awake()
-    {
-        player = GameObject.FindWithTag("Player");
-        playerHealth = player.GetComponent<PlayerHealth>();
-        meshAgent = GetComponent<NavMeshAgent>();
-        meshAgent.stoppingDistance = 1.1f; //Stop enemy pushing player
-    }
-
-    void Update()
-    {
-        meshAgent.destination = player.transform.position; 
-    }
-   
-}
