+++ Survival-Game/Assets/Scripts/Enemy/EnemySpawner.cs
@@ -4,7 +4,8 @@ using UnityEngine;
 
 public class EnemySpawner : MonoBehaviour
 {
-    [SerializeField] GameObject enemyPrefab;
+    [SerializeField] GameObject normalEnemyPrefab;
+    [SerializeField] GameObject ExplodingEnemyPrefab;
 
     [SerializeField] float spawnRate = 10.0f;
 
@@ -16,7 +17,16 @@ public class EnemySpawner : MonoBehaviour
     {
         while (true)
         {
-            Instantiate(enemyPrefab, transform.position, Quaternion.identity);
+            float randomNumber = Random.Range(0, 10.0f); //more likely for normal enemy
+            if (randomNumber > 3.0f)
+            {
+                Instantiate(normalEnemyPrefab, transform.position, Quaternion.identity);
+            }
+            else
+            {
+                Instantiate(ExplodingEnemyPrefab, transform.position, Quaternion.identity);
+            }
+           
             yield return new WaitForSeconds(spawnRate); 
         }
     }
