+++ Survival-Game/Assets/Scripts/Enemy/ExplodingRobotAI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a5042d687f2c4984586962fb5e425a6f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
